requirements.txt
fastapi
uvicorn
python-dotenv
openai
qdrant-client
langchain
langchain-community
langchain-openai
tiktoken
